pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "esgwater"
version = "0.1.0"
description = "ESG + Water scarcity scoring"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.2",
  "numpy>=1.26",
  "scikit-learn>=1.4",
  "scipy>=1.12",
  "pydantic>=2.8",
  "pyyaml>=6.0",
  "requests>=2.32",
  "click>=8.0",
  "rich>=13.7",
  "matplotlib>=3.8",
  "plotly>=5.22",
  "shap>=0.46",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
esgwater = "esgwater.cli:cli"